Add files via upload
Task 1- Student Management System
The Task was to Create a Student class to represent individual students. includes attributes such as roll no. ,name , grade and any other relevant details of student. Also includes an method to add a student, remove a student and search for a student and display all students.